requirements.txt
Pillow==10.1.0
Pillow==10.1.0
pynput==1.7.6
